Attempt at making landing over LDA distance calculator

+++ src/main/java/uk/ac/soton/comp2211/team33/models/Obstacle.java
@@ -14,11 +14,11 @@ public class Obstacle {
     this.distanceThresh = distanceThresh;
   }
 
-  public int getHeight() {
+  public double getHeight() {
     return height;
   }
 
-  public int getLength() {
+  public double getLength() {
     return length;
   }
